ensure genesis block generation event is properly broadcast to network
