+++ crates/node/src/network/handler.rs
@@ -43,10 +43,14 @@ impl Handler<EventMessage> for NetworkModule {
                     .await
                     .map_err(|err| TheaterError::Other(err.to_string()))?;
             },
-
-            Event::QuorumMembershipAssigmentCreated(assigned_membership) => {
-                self.notify_quorum_membership_assignment(assigned_membership)
+            Event::QuorumMembershipAssigmentsCreated(assigments) => {
+                self.notify_quorum_membership_assignments(assigments)
                     .await?;
+
+                // for assigned_membership in assigments {
+                //     self.notify_quorum_membership_assignment(assigned_membership)
+                //         .await?;
+                // }
             },
 
             Event::ClaimCreated(claim) => {
