+++ crates/node/src/runtime/mod.rs
@@ -1,122 +1,14 @@
-use events::{Event, EventPublisher, EventRouter, Vote};
-use primitives::{JSON_RPC_API_TOPIC_STR, NETWORK_TOPIC_STR, RUNTIME_TOPIC_STR};
-use telemetry::info;
-use vrrb_config::NodeConfig;
-
-use crate::{
-    api::setup_rpc_api_server,
-    component::NodeRuntimeComponentConfig,
-    indexer_module::setup_indexer_module,
-    network::{NetworkModule, NetworkModuleComponentConfig},
-    node_runtime::NodeRuntime,
-    result::Result,
-    ui::setup_node_gui,
-    RuntimeComponent, RuntimeComponentManager,
-};
-
 pub mod component;
 pub mod handler_helpers;
 pub mod node_runtime;
 pub mod node_runtime_handler;
+mod setup;
 
 pub use handler_helpers::*;
-
-pub const PULL_TXN_BATCH_SIZE: usize = 100;
-
-pub async fn setup_runtime_components(
-    original_config: &NodeConfig,
-    router: &EventRouter,
-    events_tx: EventPublisher,
-) -> Result<(RuntimeComponentManager, NodeConfig)> {
-    let mut config = original_config.clone();
-
-    let runtime_events_rx = router.subscribe(Some(RUNTIME_TOPIC_STR.into()))?;
-    let network_events_rx = router.subscribe(Some(NETWORK_TOPIC_STR.into()))?;
-    let jsonrpc_events_rx = router.subscribe(Some(JSON_RPC_API_TOPIC_STR.into()))?;
-    let indexer_events_rx = router.subscribe(None)?;
-
-    let mut runtime_manager = RuntimeComponentManager::new();
-
-    let node_runtime_component_handle = NodeRuntime::setup(NodeRuntimeComponentConfig {
-        config: config.clone(),
-        events_tx: events_tx.clone(),
-        events_rx: runtime_events_rx,
-    })
-    .await?;
-
-    let handle_data = node_runtime_component_handle.data();
-
-    let node_config = handle_data.node_config.clone();
-
-    config = node_config;
-
-    let mempool_read_handle_factory = handle_data.mempool_read_handle_factory;
-    let state_read_handle = handle_data.state_read_handle;
-
-    runtime_manager.register_component(
-        node_runtime_component_handle.label(),
-        node_runtime_component_handle.handle(),
-    );
-
-    let network_component_handle = NetworkModule::setup(NetworkModuleComponentConfig {
-        config: config.clone(),
-        node_id: config.id.clone(),
-        events_tx: events_tx.clone(),
-        network_events_rx,
-        vrrbdb_read_handle: state_read_handle.clone(),
-        bootstrap_quorum_config: config.bootstrap_quorum_config.clone(),
-        membership_config: config.quorum_config.clone(),
-        validator_public_key: config.keypair.validator_public_key_owned(),
-    })
-    .await?;
-
-    let resolved_network_data = network_component_handle.data();
-    let network_component_handle_label = network_component_handle.label();
-
-    runtime_manager.register_component(
-        network_component_handle_label,
-        network_component_handle.handle(),
-    );
-
-    config.kademlia_peer_id = Some(resolved_network_data.kademlia_peer_id);
-    config.udp_gossip_address = resolved_network_data.resolved_udp_gossip_address;
-    config.raptorq_gossip_address = resolved_network_data.resolved_raptorq_gossip_address;
-    config.kademlia_liveness_address = resolved_network_data.resolved_kademlia_liveness_address;
-
-    let (jsonrpc_server_handle, resolved_jsonrpc_server_addr) = setup_rpc_api_server(
-        &config,
-        events_tx.clone(),
-        state_read_handle.clone(),
-        mempool_read_handle_factory.clone(),
-        jsonrpc_events_rx,
-    )
-    .await?;
-
-    config.jsonrpc_server_address = resolved_jsonrpc_server_addr;
-
-    info!("JSON-RPC server address: {}", config.jsonrpc_server_address);
-
-    runtime_manager.register_component("API".to_string(), jsonrpc_server_handle);
-
-    if config.enable_block_indexing {
-        let handle = setup_indexer_module(&config, indexer_events_rx, mempool_read_handle_factory)?;
-        // TODO: udpate this to return the proper component handle type
-        // indexer_handle = Some(handle);
-        // TODO: register indexer module handle
-    }
-
-    let mut node_gui_handle = None;
-    if config.gui {
-        node_gui_handle = setup_node_gui(&config).await?;
-        info!("Node UI started");
-    }
-
-    Ok((runtime_manager, config))
-}
+pub use setup::*;
 
 #[cfg(test)]
 mod tests {
-    use std::collections::HashMap;
 
     use crate::node_runtime::NodeRuntime;
     use crate::test_utils::{
@@ -126,13 +18,11 @@ mod tests {
         setup_network, setup_whitelisted_nodes,
     };
     use crate::NodeError;
-    use block::{Block, GenesisConfig, GenesisReceiver};
+    use block::{Block, GenesisReceiver};
     use events::{AssignedQuorumMembership, PeerData, Vote, DEFAULT_BUFFER};
     use primitives::{generate_account_keypair, Address, NodeId, NodeType, QuorumKind};
     use storage::storage_utils::remove_vrrb_data_dir;
-    use tracing::instrument::WithSubscriber;
-    use vrrb_config::QuorumMember;
-    use vrrb_core::account::{self, Account, AccountField};
+    use vrrb_core::account::{Account, AccountField};
     use vrrb_core::transactions::Transaction;
 
     #[tokio::test]
@@ -220,6 +110,7 @@ mod tests {
         let (_node_0, farmers, harvesters, _miners) = setup_network(8).await;
         let mut validators = farmers.clone();
         validators.extend(harvesters.clone().into_iter());
+
         for (farmer_id, farmer) in farmers.iter() {
             for (validator_id, member) in validators.iter() {
                 if validator_id == farmer_id {
