+++ crates/node/src/test_utils.rs
@@ -22,8 +22,8 @@ use crate::{
 use events::{AssignedQuorumMembership, EventPublisher, PeerData, DEFAULT_BUFFER};
 pub use miner::test_helpers::{create_address, create_claim, create_miner};
 use primitives::{
-    generate_account_keypair, Address, KademliaPeerId, NodeId, NodeType, QuorumKind, RawSignature,
-    Round, Signature,
+    generate_account_keypair, Address, KademliaPeerId, NodeId, NodeType, QuorumKind, Round,
+    Signature,
 };
 use rand::{seq::SliceRandom, thread_rng};
 use secp256k1::{Message, PublicKey, SecretKey};
@@ -604,6 +604,10 @@ impl DataStore<MockStateReader> for MockStateStore {
 
 /// Creates `n` Node instances that make up a network.
 pub async fn create_test_network(n: u16) -> Vec<Node> {
+    create_test_network_from_config(n, None).await
+}
+
+pub async fn create_test_network_from_config(n: u16, base_config: Option<NodeConfig>) -> Vec<Node> {
     let validator_count = (n as f64 * 0.8).ceil() as usize;
     let miner_count = n as usize - validator_count;
 
@@ -615,8 +619,6 @@ pub async fn create_test_network(n: u16) -> Vec<Node> {
         let raptor_port: u16 = 12000 + i;
         let kademlia_port: u16 = 13000 + i;
 
-        // let threshold_sk = ValidatorSecretKey::random();
-        // let validator_public_key = threshold_sk.public_key();
         let keypair = Keypair::random();
         let validator_public_key = keypair.miner_public_key_owned();
 
@@ -638,6 +640,11 @@ pub async fn create_test_network(n: u16) -> Vec<Node> {
         quorum_members.insert(node_id, member);
     }
 
+    let whitelisted_nodes = quorum_members
+        .values()
+        .cloned()
+        .collect::<Vec<QuorumMember>>();
+
     let bootstrap_quorum_config = BootstrapQuorumConfig {
         membership_config: QuorumMembershipConfig {
             quorum_members: quorum_members.clone(),
@@ -646,10 +653,17 @@ pub async fn create_test_network(n: u16) -> Vec<Node> {
         genesis_transaction_threshold: (n / 2) as u64,
     };
 
-    let mut config = create_mock_full_node_config();
+    let mut config = if let Some(base_config) = base_config.clone() {
+        base_config
+    } else {
+        create_mock_full_node_config()
+    };
+
     config.id = String::from("node-0");
 
     config.bootstrap_quorum_config = Some(bootstrap_quorum_config.clone());
+
+    config.whitelisted_nodes = whitelisted_nodes.clone();
 
     let node_0 = Node::start(config).await.unwrap();
 
@@ -681,6 +695,7 @@ pub async fn create_test_network(n: u16) -> Vec<Node> {
         config.raptorq_gossip_address = quorum_config.raptorq_gossip_address;
         config.udp_gossip_address = quorum_config.udp_gossip_address;
         config.kademlia_peer_id = Some(quorum_config.kademlia_peer_id);
+        config.whitelisted_nodes = whitelisted_nodes.clone();
 
         let node = Node::start(config).await.unwrap();
         nodes.push(node);
@@ -699,6 +714,7 @@ pub async fn create_test_network(n: u16) -> Vec<Node> {
         miner_config.raptorq_gossip_address = quorum_config.raptorq_gossip_address;
         miner_config.udp_gossip_address = quorum_config.udp_gossip_address;
         miner_config.kademlia_peer_id = Some(quorum_config.kademlia_peer_id);
+        miner_config.whitelisted_nodes = whitelisted_nodes.clone();
 
         let miner_node = Node::start(miner_config).await.unwrap();
 
@@ -706,6 +722,10 @@ pub async fn create_test_network(n: u16) -> Vec<Node> {
     }
 
     nodes
+}
+
+fn setup_kademlia_config() {
+    //
 }
 
 /// Creates n NodeRuntimes to simulate networks
